move client cert functions into separate resource
