+++ graylog2-server/src/main/java/org/graylog2/rest/resources/RestResourcesModule.java
@@ -18,6 +18,7 @@ package org.graylog2.rest.resources;
 
 import org.graylog.plugins.views.search.engine.monitoring.data.histogram.rest.HistogramResponseWriter;
 import org.graylog.plugins.views.storage.migration.RemoteReindexResource;
+import org.graylog.security.rest.CAClientResource;
 import org.graylog.security.rest.CAResource;
 import org.graylog.security.rest.CertificateRenewalResource;
 import org.graylog2.Configuration;
@@ -162,6 +163,7 @@ public class RestResourcesModule extends Graylog2Module {
         addSystemRestResource(DataNodeManagementResource.class);
         addSystemRestResource(RemoteReindexResource.class);
         addSystemRestResource(CAResource.class);
+        addSystemRestResource(CAClientResource.class);
     }
 
     private void addDebugResources() {
